new Data structure
basically got rid of how the current task/project works now it only holds ID and then uses that ID to search for it in the data. so i dont have to do that weird data matching thing. did some slight work on ui inputs. also changed all functions so it works with new data structure